// types/pages/landing.d.ts
import { Header } from '@/types/blocks/header';
import { Hero } from '@/types/blocks/hero';
import { Section } from '@/types/blocks/section';
import { Footer } from '@/types/blocks/footer';

export interface Meta {
  title: string;
  description: string;
}

export interface LandingPage {
  meta: Meta;
  header?: Header;
  hero?: Hero;
  branding?: Section;
  introduce?: Section;
  benefit?: Section;
  usage?: Section;
  feature?: Section;
  showcase?: Section;
  stats?: Section;
  pricing?: Pricing;
  testimonial?: Section;
  faq?: Section;
  cta?: Section;
  footer?: Footer;
}